// src/models/Menu/IMenu.tsx
import { NavigateFunction } from "react-router-dom";

export interface IMenuProps {
  switchPaddingMenu: CallableFunction;
  width: number;
  widthLogoIcon: CallableFunction;
  navigate: NavigateFunction;
  openMenuMobile: any;
  setOpenMenuMobile: any;
  headerActive: any;
  location: any;
  LogoBdiscomPrimary: any;
  LogoBdiscomWhite: any;
  itemMenu: any;
}